renamed the project name from HelloWorld to KetchPrefCenterSample; renamed the package name to com.ketch.sample.prefcenter; implemented AdvertisingId getting;
