+++ ketch-android/Android Ketch Pref Center using Kotlin/app/build.gradle
@@ -8,7 +8,7 @@ android {
     compileSdk 33
 
     defaultConfig {
-        applicationId "com.example.helloworld"
+        applicationId "com.ketch.sample.prefcenter"
         minSdk 21
         targetSdk 33
         versionCode 1
@@ -43,6 +43,9 @@ dependencies {
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     implementation 'androidx.webkit:webkit:1.5.0'
     implementation 'com.google.code.gson:gson:2.9.1'
+
+    implementation 'com.google.android.gms:play-services-base:18.1.0'
+    implementation 'com.google.android.gms:play-services-ads-identifier:18.0.1'
 
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
