+++ ketch-android/Android Ketch Pref Center using Kotlin/app/src/androidTest/java/com/ketch/sample/prefcenter/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.helloworld
+package com.ketch.sample.prefcenter
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.helloworld", appContext.packageName)
+        assertEquals("com.ketch.sample.prefcenter", appContext.packageName)
     }
 }
