+++ ketch-android/Android Ketch Pref Center using Kotlin/app/src/main/AndroidManifest.xml
@@ -1,8 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.example.helloworld">
-    <uses-permission android:name="android.permission.INTERNET"/>
+    package="com.ketch.sample.prefcenter">
+
+    <uses-permission android:name="android.permission.INTERNET" />
+
     <application
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
@@ -11,7 +13,7 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.HelloWorld"
+        android:theme="@style/Theme.KetchSample"
         tools:targetApi="31">
         <activity
             android:name=".MainActivity"
@@ -22,7 +24,8 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".KetchPrefCenter"></activity>
+
+        <activity android:name=".KetchPrefCenter"/>
     </application>
 
 </manifest>
