+++ ketch-android/Android Ketch Pref Center using Kotlin/app/src/main/java/com/ketch/sample/prefcenter/Identity.kt
@@ -1,4 +1,4 @@
-package com.example.helloworld
+package com.ketch.sample.prefcenter
 
 import android.os.Parcelable
 import kotlinx.android.parcel.Parcelize
