+++ ketch-android/Android Ketch Pref Center using Kotlin/app/src/main/res/layout/activity_main.xml
@@ -13,7 +13,7 @@
         android:layout_marginStart="176dp"
         android:layout_marginTop="360dp"
         android:layout_marginEnd="177dp"
-        android:text="Hello World!"
+        android:text="@string/app_name"
         android:textSize="34sp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.496"
@@ -27,7 +27,7 @@
         android:layout_marginStart="159dp"
         android:layout_marginTop="76dp"
         android:layout_marginEnd="158dp"
-        android:text="Preference Center"
+        android:text="@string/show"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.504"
         app:layout_constraintStart_toStartOf="parent"
