+++ ketch-android/Android Ketch Pref Center using Kotlin/app/src/main/res/values/strings.xml
@@ -1,6 +1,6 @@
 <resources>
-    <string name="app_name">Hello World</string>
-    <string name="title_activity_ketch_pref_center">KetchPrefCenter</string>
-    <string name="dummy_button">Dummy Button</string>
-    <string name="dummy_content">DUMMY\nCONTENT</string>
+    <string name="app_name">Ketch PrefCenter</string>
+    <string name="show">Show</string>
+
+    <string name="cannot_get_advertising_id_toast">Cannot get AdvertisingId</string>
 </resources>
