+++ ketch-android/Android Ketch Pref Center using Kotlin/app/src/main/res/values/styles.xml
@@ -1,10 +1,10 @@
 <resources>
 
-    <style name="Widget.Theme.HelloWorld.ActionBar.Fullscreen" parent="Widget.AppCompat.ActionBar">
+    <style name="Widget.Theme.KetchSample.ActionBar.Fullscreen" parent="Widget.AppCompat.ActionBar">
         <item name="android:background">@color/black_overlay</item>
     </style>
 
-    <style name="Widget.Theme.HelloWorld.ButtonBar.Fullscreen" parent="">
+    <style name="Widget.Theme.KetchSample.ButtonBar.Fullscreen" parent="">
         <item name="android:background">@color/black_overlay</item>
         <item name="android:buttonBarStyle">?android:attr/buttonBarStyle</item>
     </style>
