+++ ketch-android/Android Ketch Pref Center using Kotlin/app/src/test/java/com/ketch/sample/prefcenter/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.example.helloworld
+package com.ketch.sample.prefcenter
 
 import org.junit.Test
