+++ ketch-android/Android Ketch Pref Center using Kotlin/settings.gradle
@@ -12,5 +12,5 @@ dependencyResolutionManagement {
         mavenCentral()
     }
 }
-rootProject.name = "Hello World"
+rootProject.name = "KetchPrefCenterSample"
 include ':app'
